started replacing Context with Context2 (non-compiling)
